fix error

+++ TrainModel.py
@@ -72,7 +72,7 @@ def run_training():
     sketchs_placeholder = tf.placeholder(tf.float32)
     images_neg_placeholder = tf.placeholder(tf.float32)
     images_pos_placeholder = tf.placeholder(tf.float32)
-    keep_prob = tf.placeholder(tf.float64)
+    keep_prob = tf.placeholder(tf.float32)
 
     # Three Branch Net
     sketch_dense = SketchNet(sketchs_placeholder, _weights = image_weights, _biases = image_biases, dropout_prob = keep_prob)
